+++ ArtifactOfLimit/Thunderstore/CHANGELOG.md
@@ -1,4 +1,9 @@
 <details>
+<summary>1.0.3 </summary>
+
+* Fixed game hanging if you go to the next stage with Artifact disabled and ProperSave enabled.
+</details>
+<details>
 <summary>1.0.2 </summary>
 
 * Recompiled for AC.